Move UserBuilder to its own file
