+++ tripservice/UserBuilder.py
@@ -1,0 +1,27 @@
+from User import User
+
+class UserBuilder:
+    def aUser():
+        return UserBuilder()
+
+    def friendsWith(self, friends):
+        self.friends = friends
+        return self
+
+    def withTrips(self, trips):
+        self.trips = trips
+        return self
+
+    def build(self):
+        user = User()
+        self._addTripsTo(user)
+        self._addFriendsTo(user)
+        return user
+
+    def _addFriendsTo(self, user):
+        for friend in self.friends:
+            user.addFriend(friend)
+
+    def _addTripsTo(self, user):
+        for trip in self.trips:
+            user.addTrip(trip)
