+++ tripservice/test_TripService.py
@@ -3,6 +3,7 @@ from unittest import TestCase
 from DependendClassCallDuringUnitTestException import *
 from Trip import *
 from TripService import TripService
+from UserBuilder import UserBuilder
 from UserNotLoggedInException import UserNotLoggedInException
 from User import User
 
@@ -50,30 +51,3 @@ class TestableTripService(TripService):
 
     def findTripsByUser(self, user):
         return user.getTrips()
-
-
-class UserBuilder:
-    def aUser():
-        return UserBuilder()
-
-    def friendsWith(self, friends):
-        self.friends = friends
-        return self
-
-    def withTrips(self, trips):
-        self.trips = trips
-        return self
-
-    def build(self):
-        user = User()
-        self._addTripsTo(user)
-        self._addFriendsTo(user)
-        return user
-
-    def _addFriendsTo(self, user):
-        for friend in self.friends:
-            user.addFriend(friend)
-
-    def _addTripsTo(self, user):
-        for trip in self.trips:
-            user.addTrip(trip)
